+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.0 - _10.09.2024_
+
+**Minor Changes**
+
+- Changed a few event message
+- Added/Implemented event "reopened issue"
+
 ## 1.0.0 - _24.08.2024_
 
 **Initial Release**